% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dependencies.r
\name{install_dependencies}
\alias{install_dependencies}
\title{Install stiltread dependencies}
\usage{
install_dependencies(...)
}
\description{
Install stiltread dependencies
}
\author{
Ben Fasoli

Run after installing stiltread to set up dependencies on system
}
