% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_stiltread.r
\name{install_stiltread}
\alias{install_stiltread}
\title{Install stiltread dependencies}
\usage{
install_stiltread(...)
}
\description{
Install stiltread dependencies
}
\author{
Ben Fasoli

Run after installing stiltread to set up dependencies on system
}
