% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_met.r
\name{read_met}
\alias{read_met}
\title{Read ARL packed meteorological data file layer}
\usage{
read_met(path, var)
}
\arguments{
\item{path}{to ARL packed file}

\item{var}{four digit character code of variable. Examples include SHGT PRSS
MSLP TPP1 DIFR T02M DP2M PBLH DSWF ULWF U10M V10M VSBY CSNO CICE CFZR CRAI
LIB4 CAPE CINH TCLD REFC CLDB CLDT}
}
\description{
Read ARL packed meteorological data file layer
}
\author{
Ben Fasoli

Extracts the given variable from file and returns a raster with associated
projection metadata
}
