% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_met_header.r
\name{read_met_header}
\alias{read_met_header}
\title{Read ARL packed meteorological data file header}
\usage{
read_met_header(path)
}
\arguments{
\item{path}{to ARL packed file}
}
\description{
Read ARL packed meteorological data file header
}
\author{
Ben Fasoli

Extracts metadata from the meteorological data file header
}
