% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_met_wind.r
\name{read_met_wind}
\alias{read_met_wind}
\title{Read ARL packed meteorological data file winds}
\usage{
read_met_wind(path, yy, mm, dd, hh, lvl, wnd_warning = T)
}
\arguments{
\item{path}{to ARL packed file}

\item{yy}{desired record year}

\item{mm}{desired record month of year}

\item{dd}{desired record day of month}

\item{hh}{desired record hour of day}

\item{lvl}{desired record vertical level (0:nz)}
}
\value{
rasterStack with u and v rasterLayers projected into '+proj=longlat'
}
\description{
Read ARL packed meteorological data file winds
}
\author{
Ben Fasoli

Extracts the winds at given level from file. For the surface level (0), 10m
u and v winds are reported.
}
