% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_met.r
\name{read_met}
\alias{read_met}
\title{Read ARL packed meteorological data file layer}
\usage{
read_met(path, var, yy, mm, dd, hh, lvl, wnd_warning = T)
}
\arguments{
\item{path}{to ARL packed file}

\item{var}{four digit character code of variable. Examples include SHGT PRSS
MSLP TPP1 DIFR T02M DP2M PBLH DSWF ULWF U10M V10M VSBY CSNO CICE CFZR CRAI
LIB4 CAPE CINH TCLD REFC CLDB CLDT}

\item{yy}{desired record year}

\item{mm}{desired record month of year}

\item{dd}{desired record day of month}

\item{hh}{desired record hour of day}

\item{lvl}{desired record vertical level (0:nz)}

\item{wnd_warning}{report grid orientation warning when querying wind data}
}
\value{
rasterLayer with associated projection metadata
}
\description{
Read ARL packed meteorological data file layer
}
\author{
Ben Fasoli

Extracts the given variable from file
}
